Changed received_file_size to record the progress
The received_file_size records now how much data has been received (ignoring duplicates) rather than the offset of the last data received.

This is done by implementing logic in the segments::update_segments to return the amount of new data.

The offset of the last data received (which is the previous meaning of the received_file_size) is the end of the last segment and it does not need to be stored separately.

Fixes #42