Added a delay parameter to the NAK procedure configuration
The delay parameter helps when the PDUs arrive out of order. If the NAKs are sent immediately (in case of the Nak immediate) procedure, then many will be sent for data that is just abbout to arrive. For deferred procedure the same happens when EOF is received.
The delay parameter causes the checking for gaps to be performed a little later when the PDUs in transit had a chance to be processed.

Fixes #48